move generation to functions
